go: more region-smart replication
When syncing a file to multiple devices in a remote region, only
actually sync with one of them. For handoff replication, do an
in-sync check on the others so we can still remove objects that are
fully synced.  For non-handoff data, just send a ping to keep the
connection alive.

Plus a little bit of cleanup.

This is similar to how swift replication already works.

Change-Id: I190f97e7ad8249af6f7ecd44977a2e4d6912bb90